fix: prevent out-of-scope repair deep-link create
