+++ src/app/(app)/repair-requests/__tests__/useRepairRequestsDeepLink.test.ts
@@ -34,8 +34,10 @@ vi.mock('@/lib/rpc-client', () => ({
 }))
 
 // ── Import AFTER mocks ────────────────────────────────────────────
-import { useRepairRequestsDeepLink } from '../_hooks/useRepairRequestsDeepLink'
 import type { UiFilters } from '@/lib/rr-prefs'
+import { registerUseRepairRequestsDeepLinkRaceCases } from './useRepairRequestsDeepLink.race-cases'
+import { registerUseRepairRequestsDeepLinkRetryCases } from './useRepairRequestsDeepLink.retry-cases'
+import { useRepairRequestsDeepLink } from '../_hooks/useRepairRequestsDeepLink'
 
 // ── Helpers ───────────────────────────────────────────────────────
 function createSearchParams(params: Record<string, string> = {}): URLSearchParams {
@@ -222,7 +224,7 @@ describe('useRepairRequestsDeepLink', () => {
     expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
   })
 
-  it('degrades gracefully when action=create has an unresolved equipmentId', async () => {
+  it('does not open the create sheet when action=create has an unresolved equipmentId', async () => {
     mocks.callRpc
       .mockResolvedValueOnce([])
       .mockResolvedValueOnce(null)
@@ -232,11 +234,56 @@ describe('useRepairRequestsDeepLink', () => {
     renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
 
     await waitFor(() => {
-      expect(mocks.openCreateSheet).toHaveBeenCalledWith()
+      expect(mocks.toast).toHaveBeenCalledWith(
+        expect.objectContaining({
+          variant: 'destructive',
+          title: 'Lỗi',
+          description: expect.stringContaining('Không thể mở phiếu sửa chữa'),
+        })
+      )
     })
-    expect(mocks.openCreateSheet).not.toHaveBeenCalledWith(
-      expect.objectContaining({ id: 999 })
-    )
+    expect(mocks.openCreateSheet).not.toHaveBeenCalled()
+    expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
+  })
+
+  it('does not open the create sheet when equipment_get denies access', async () => {
+    mocks.callRpc
+      .mockResolvedValueOnce([])
+      .mockRejectedValueOnce({ message: 'Equipment not found or access denied' })
+
+    const sp = createSearchParams({ action: 'create', equipmentId: '999' })
+
+    renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
+
+    await waitFor(() => {
+      expect(mocks.toast).toHaveBeenCalledWith(
+        expect.objectContaining({
+          variant: 'destructive',
+          title: 'Lỗi',
+          description: expect.stringContaining('Equipment not found or access denied'),
+        })
+      )
+    })
+    expect(mocks.openCreateSheet).not.toHaveBeenCalled()
+    expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
+  })
+
+  it('does not treat an invalid equipmentId as a blank create intent', async () => {
+    const sp = createSearchParams({ action: 'create', equipmentId: 'abc' })
+
+    renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
+
+    await waitFor(() => {
+      expect(mocks.toast).toHaveBeenCalledWith(
+        expect.objectContaining({
+          variant: 'destructive',
+          title: 'Lỗi',
+          description: expect.stringContaining('Không thể mở phiếu sửa chữa'),
+        })
+      )
+    })
+    expect(mocks.openCreateSheet).not.toHaveBeenCalled()
+    expect(mocks.callRpc).toHaveBeenCalledTimes(1)
     expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
   })
 
@@ -250,340 +297,17 @@ describe('useRepairRequestsDeepLink', () => {
     expect(source).not.toContain("searchParams.get('action') !== 'create'")
   })
 
-  // ── Race condition tests ────────────────────────────────────────
-  // These use deferred promises to control timing between equipment_list
-  // and equipment_get, exposing the race in the current implementation.
+  registerUseRepairRequestsDeepLinkRaceCases({
+    useRepairRequestsDeepLink,
+    mocks,
+    createDefaultOptions,
+    createSearchParams,
+  })
 
-  function createDeferred<T>() {
-    let resolve!: (value: T) => void
-    let reject!: (reason?: unknown) => void
-    const promise = new Promise<T>((res, rej) => { resolve = res; reject = rej })
-    return { promise, resolve, reject }
-  }
-
-  const VALID_EQUIPMENT = {
-    id: 42,
-    ma_thiet_bi: 'TB042',
-    ten_thiet_bi: 'Máy B',
-    khoa_phong_quan_ly: 'Khoa 2',
-  }
-
-  describe('race: equipment resolution timing', () => {
-    it('waits for targeted equipment_get before opening sheet when list settles first', async () => {
-      const equipmentGetDeferred = createDeferred<typeof VALID_EQUIPMENT | null>()
-
-      mocks.callRpc
-        .mockResolvedValueOnce([])              // equipment_list: resolves immediately
-        .mockReturnValueOnce(equipmentGetDeferred.promise) // equipment_get: deferred
-
-      const sp = createSearchParams({ action: 'create', equipmentId: '42' })
-      renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
-
-      // Wait for list to settle
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(2)
-      })
-
-      // Sheet must NOT have opened yet while equipment_get is still pending
-      expect(mocks.openCreateSheet).not.toHaveBeenCalled()
-
-      // Now resolve equipment_get
-      await act(async () => {
-        equipmentGetDeferred.resolve(VALID_EQUIPMENT)
-      })
-
-      // Sheet should now open WITH equipment
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith(
-          expect.objectContaining({ id: 42, ma_thiet_bi: 'TB042' })
-        )
-      })
-      expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
-    })
-
-    it('opens with prefill when targeted equipment_get resolves before list is useful', async () => {
-      const listDeferred = createDeferred<Array<typeof VALID_EQUIPMENT>>()
-
-      mocks.callRpc
-        .mockReturnValueOnce(listDeferred.promise)        // equipment_list: deferred (slow)
-        .mockResolvedValueOnce(VALID_EQUIPMENT)            // equipment_get: resolves immediately
-
-      const sp = createSearchParams({ action: 'create', equipmentId: '42' })
-      renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
-
-      // equipment_get resolves quickly; sheet should open with prefill
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith(
-          expect.objectContaining({ id: 42, ma_thiet_bi: 'TB042' })
-        )
-      })
-
-      // Cleanup: resolve the list to avoid dangling promise
-      await act(async () => {
-        listDeferred.resolve([])
-      })
-    })
-
-    it('does not reopen the create sheet when the initial list settles after the intent is consumed', async () => {
-      const listDeferred = createDeferred<Array<typeof VALID_EQUIPMENT>>()
-
-      mocks.callRpc
-        .mockReturnValueOnce(listDeferred.promise)
-        .mockResolvedValueOnce(VALID_EQUIPMENT)
-
-      const sp = createSearchParams({ action: 'create', equipmentId: '42' })
-      const opts = createDefaultOptions(sp)
-      renderHook(() => useRepairRequestsDeepLink(opts))
-
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledTimes(1)
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith(
-          expect.objectContaining({ id: 42, ma_thiet_bi: 'TB042' })
-        )
-      })
-
-      await act(async () => {
-        listDeferred.resolve([VALID_EQUIPMENT])
-      })
-
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledTimes(1)
-      })
-    })
-
-    it('opens blank sheet only after equipment_get reaches terminal missing state', async () => {
-      const equipmentGetDeferred = createDeferred<null>()
-
-      mocks.callRpc
-        .mockResolvedValueOnce([])                          // equipment_list: immediate
-        .mockReturnValueOnce(equipmentGetDeferred.promise)  // equipment_get: deferred
-
-      const sp = createSearchParams({ action: 'create', equipmentId: '999' })
-      renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
-
-      // Wait for list
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(2)
-      })
-
-      // Sheet must NOT open while equipment_get is pending
-      expect(mocks.openCreateSheet).not.toHaveBeenCalled()
-
-      // Resolve as missing
-      await act(async () => {
-        equipmentGetDeferred.resolve(null)
-      })
-
-      // Now sheet should open blank (graceful degradation)
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith()
-      })
-      expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
-    })
-
-    it('waits for the latest equipmentId when the URL changes mid-flight', async () => {
-      const firstEquipmentDeferred = createDeferred<typeof VALID_EQUIPMENT | null>()
-      const secondEquipmentDeferred = createDeferred<{
-        id: 99
-        ma_thiet_bi: string
-        ten_thiet_bi: string
-        khoa_phong_quan_ly: string
-      } | null>()
-      const nextEquipment = {
-        id: 99 as const,
-        ma_thiet_bi: 'TB099',
-        ten_thiet_bi: 'Máy C',
-        khoa_phong_quan_ly: 'Khoa 3',
-      }
-      const baseOpts = createDefaultOptions()
-
-      mocks.callRpc
-        .mockResolvedValueOnce([])
-        .mockReturnValueOnce(firstEquipmentDeferred.promise)
-        .mockReturnValueOnce(secondEquipmentDeferred.promise)
-
-      const { rerender } = renderHook(
-        ({ currentSearchParams }) => useRepairRequestsDeepLink({
-          ...baseOpts,
-          searchParams: currentSearchParams,
-        }),
-        {
-          initialProps: {
-            currentSearchParams: createSearchParams({ action: 'create', equipmentId: '42' }),
-          },
-        },
-      )
-
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(2)
-      })
-
-      act(() => {
-        rerender({
-          currentSearchParams: createSearchParams({ action: 'create', equipmentId: '99' }),
-        })
-      })
-
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(3)
-      })
-
-      await act(async () => {
-        firstEquipmentDeferred.resolve(VALID_EQUIPMENT)
-        await Promise.resolve()
-      })
-
-      expect(mocks.openCreateSheet).not.toHaveBeenCalled()
-      expect(mocks.routerReplace).not.toHaveBeenCalled()
-
-      await act(async () => {
-        secondEquipmentDeferred.resolve(nextEquipment)
-      })
-
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith(
-          expect.objectContaining({ id: 99, ma_thiet_bi: 'TB099' })
-        )
-      })
-      expect(mocks.openCreateSheet).not.toHaveBeenCalledWith(
-        expect.objectContaining({ id: 42 })
-      )
-      expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
-    })
-
-    it('keeps isEquipmentFetchPending true while a newer equipment fetch is still in flight', async () => {
-      const firstEquipmentDeferred = createDeferred<typeof VALID_EQUIPMENT | null>()
-      const secondEquipmentDeferred = createDeferred<{
-        id: 99
-        ma_thiet_bi: string
-        ten_thiet_bi: string
-        khoa_phong_quan_ly: string
-      } | null>()
-      const nextEquipment = {
-        id: 99 as const,
-        ma_thiet_bi: 'TB099',
-        ten_thiet_bi: 'Máy C',
-        khoa_phong_quan_ly: 'Khoa 3',
-      }
-      const baseOpts = createDefaultOptions()
-
-      mocks.callRpc
-        .mockResolvedValueOnce([])
-        .mockReturnValueOnce(firstEquipmentDeferred.promise)
-        .mockReturnValueOnce(secondEquipmentDeferred.promise)
-
-      const { result, rerender } = renderHook(
-        ({ currentSearchParams }) => useRepairRequestsDeepLink({
-          ...baseOpts,
-          searchParams: currentSearchParams,
-        }),
-        {
-          initialProps: {
-            currentSearchParams: createSearchParams({ action: 'create', equipmentId: '42' }),
-          },
-        },
-      )
-
-      await waitFor(() => {
-        expect(result.current.isEquipmentFetchPending).toBe(true)
-      })
-
-      act(() => {
-        rerender({
-          currentSearchParams: createSearchParams({ action: 'create', equipmentId: '99' }),
-        })
-      })
-
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(3)
-      })
-
-      await act(async () => {
-        firstEquipmentDeferred.resolve(VALID_EQUIPMENT)
-        await Promise.resolve()
-      })
-
-      expect(result.current.isEquipmentFetchPending).toBe(true)
-
-      await act(async () => {
-        secondEquipmentDeferred.resolve(nextEquipment)
-      })
-
-      await waitFor(() => {
-        expect(result.current.isEquipmentFetchPending).toBe(false)
-      })
-    })
-
-    it('resets same-id missing resolution to pending before retrying create intent', async () => {
-      const retryEquipmentDeferred = createDeferred<typeof VALID_EQUIPMENT | null>()
-      const baseOpts = createDefaultOptions()
-
-      mocks.callRpc
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce(null)
-        .mockReturnValueOnce(retryEquipmentDeferred.promise)
-
-      const { rerender } = renderHook(
-        ({ currentSearchParams }) => useRepairRequestsDeepLink({
-          ...baseOpts,
-          searchParams: currentSearchParams,
-        }),
-        {
-          initialProps: {
-            currentSearchParams: createSearchParams({ action: 'create', equipmentId: '42' }),
-          },
-        },
-      )
-
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith()
-      })
-
-      mocks.openCreateSheet.mockClear()
-      mocks.routerReplace.mockClear()
-
-      act(() => {
-        rerender({ currentSearchParams: createSearchParams() })
-      })
-
-      act(() => {
-        rerender({
-          currentSearchParams: createSearchParams({ action: 'create', equipmentId: '42' }),
-        })
-      })
-
-      await waitFor(() => {
-        expect(mocks.callRpc).toHaveBeenCalledTimes(3)
-      })
-
-      expect(mocks.openCreateSheet).not.toHaveBeenCalled()
-      expect(mocks.routerReplace).not.toHaveBeenCalled()
-
-      await act(async () => {
-        retryEquipmentDeferred.resolve(VALID_EQUIPMENT)
-      })
-
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith(
-          expect.objectContaining({ id: 42, ma_thiet_bi: 'TB042' })
-        )
-      })
-    })
-
-    it('does not delay action=create without equipmentId due to resolution gating', async () => {
-      mocks.callRpc.mockResolvedValueOnce([]) // equipment_list
-
-      const sp = createSearchParams({ action: 'create' })
-      renderHook(() => useRepairRequestsDeepLink(createDefaultOptions(sp)))
-
-      // Should open immediately — no equipment resolution gating
-      await waitFor(() => {
-        expect(mocks.openCreateSheet).toHaveBeenCalledWith()
-      })
-
-      // Only 1 RPC call (equipment_list), no equipment_get
-      expect(mocks.callRpc).toHaveBeenCalledTimes(1)
-      expect(mocks.routerReplace).toHaveBeenCalledWith('/repair-requests', { scroll: false })
-    })
+  registerUseRepairRequestsDeepLinkRetryCases({
+    useRepairRequestsDeepLink,
+    mocks,
+    createDefaultOptions,
+    createSearchParams,
   })
 })
